Fix test build errors and move ChatDiceRollResult to its own file
